% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPP.R
\name{OPP}
\alias{OPP}
\title{Compute Orthogonal Procrustes Problem}
\usage{
OPP(X1, X2)
}
\arguments{
\item{X1}{first matrix}

\item{X2}{second matrix}
}
\value{
optimal orthogonal matrix
}
\description{
It compute the optimal orthogonal matrices that minimize the Frobenius distance between two matrices
}
\author{
Angela Andreella
}
