% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalDistance.R
\name{orthogonalDistance}
\alias{orthogonalDistance}
\title{Compute Orthogonal-based distance}
\usage{
orthogonalDistance(X1, X2)
}
\arguments{
\item{X1}{first matrix}

\item{X2}{second matrix}
}
\value{
orthogonal-based distance
}
\description{
It compute the Frobenius distance between two orthogonal
when the first one is aligned by OPP to the second one.
}
\author{
Angela Andreella
}
