% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalDistanceMatrix.R
\name{orthogonalDistanceMatrix}
\alias{orthogonalDistanceMatrix}
\title{Compute Orthogonal-based distance matrix}
\usage{
orthogonalDistanceMatrix(X)
}
\arguments{
\item{X}{array of orthogonal matrices where the first dimension is the
number of matrices to be analyzed}
}
\value{
orthogonal-based distance matrix
}
\description{
It compute the distance matrix based on distance from
\code{orthogonalDistance}
}
\author{
Angela Andreella
}
