% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesDistance.R
\name{procrustesDistance}
\alias{procrustesDistance}
\title{Compute Procrustes-based distance}
\usage{
orthogonalDistance(X1, X2)
}
\arguments{
\item{X1}{first matrix. It can be a data functionally aligned matrix or
estimated orthogonal matrix}

\item{X2}{second matrix. It can be a data functionally aligned matrix or
estimated orthogonal matrix}

\item{rotational}{specify if you want rotational-based distance or not. Default to FALSE.}
}
\value{
Procrustes-based distance
}
\description{
It compute the Frobenius distance between two orthogonal
when the first one is aligned by OPP to the second one.
}
\author{
Angela Andreella
}
