% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustesDistanceMatrix.R
\name{procrustesDistanceMatrix}
\alias{procrustesDistanceMatrix}
\title{Compute procrustes-based distance matrix}
\usage{
procrustesDistanceMatrix(X)
}
\arguments{
\item{X}{array of orthogonal or data aligned matrices where the first dimension is the
number of matrices to be analyzed}
}
\value{
procrustes-based distance matrix
}
\description{
It compute the distance matrix based on distance from
\code{procrustesDistance}
}
\author{
Angela Andreella
}
